Refactor: Tests fixtures under shared utils
